pyproject.toml
[tool.poetry]
name = "finalproject-ryzhova-dpo-nod"
version = "0.1.0"
description = "\"Currency Wallet CLI Application\""
authors = ["yozhik94 <ryzhova.ms@yandex.ru>"]
readme = "README.md"
packages = [
	{ include = "valutatrade_hub" },
]

[tool.poetry.dependencies]
python = "^3.12"
prettytable = "^3.16.0"
requests = "^2.32.5"
python-dotenv = "^1.2.1"
python-dateutil = "^2.9.0.post0"

[tool.poetry.scripts]
project = "valutatrade_hub.cli.interface:main"


[tool.poetry.group.dev.dependencies]
ruff = "^0.14.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["venv", ".venv", "dist", "__pycache__"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []